requirements.txt
qdrant-client
langchain
langchain-community
langchain-openai
langchain-qdrant
openai
python-dotenv
tiktoken
llama_index
llama-index-vector-stores-qdrant
isort
pycodestyle